Sistema categoria funcionando

+++ src/app/Http/Repositories/AddressRepository.php
@@ -8,17 +8,38 @@ use Illuminate\Support\Facades\Auth;
 class AddressRepository {
     public function getAddressByUser()
     {
-        return Address::where('user_id',Auth::id());
+        return Address::all()->where('user_id',Auth::id());
     }
 
     public function findAddressById($id)
+    {
+        $address = $this->getAddressById($id);
+        if($address->user_id == Auth::id() || Auth::user()->role == "ADMIN"){
+            return response()->json($address);
+        } else{
+            return response()->json([
+                'message' => 'Sem autorização para acessar o endereço!'
+            ]);
+        }
+    }
+
+    public function getAddressById($id)
     {
         return Address::findOrFail($id);
     }
 
     public function updateAddress(Address $address, $data)
     {
-        return $address->update($data);
+        if($address->user_id == Auth::id() || Auth::user()->role == "ADMIN"){
+            $address->update($data);
+            return response()->json([
+                'message' => 'Endereço atualizado com sucesso!'
+            ]);
+        } else{
+            return response()->json([
+                'message' => 'Sem autorização para atualizar o endereço!'
+            ]);
+        }
     }
 
     public function storeAddress($data)
@@ -34,8 +55,18 @@ class AddressRepository {
         ]);
     }
 
-    public function deleteAddress(Address $address)
+    public function deleteAddress($id)
     {
-        $address->delete();
+        $address = $this->getAddressById($id);
+        if($address->user_id == Auth::id() || Auth::user()->role == "ADMIN") {
+            $address->delete();
+            return response()->json([
+                'message' => 'Endereço deletado com sucesso!'
+            ]);
+        } else {
+            return response()->json([
+                'message' => 'Sem autorização para excluir o endereço!'
+            ]);
+        }
     }
 }
